+++ Cargo.toml
@@ -14,6 +14,7 @@ members = [
   "crates/primitives/sequencer-address",
   "crates/primitives/storage",
   "crates/primitives/commitments",
+  "crates/primitives/chain-id",
   "crates/client/block-proposer",
   "crates/client/db",
   "crates/client/rpc-core",
@@ -130,6 +131,7 @@ mp-state = { path = "crates/primitives/state", default-features = false }
 mp-storage = { path = "crates/primitives/storage", default-features = false }
 mp-transactions = { path = "crates/primitives/transactions", default-features = false }
 mp-commitments = { path = "crates/primitives/commitments", default-features = false }
+mp-chain-id = { path = "crates/primitives/chain-id", default-features = false }
 
 # Madara client
 mc-mapping-sync = { path = "crates/client/mapping-sync" }
