+++ crates/runtime/Cargo.toml
@@ -54,6 +54,7 @@ pallet-starknet = { workspace = true }
 mp-felt = { workspace = true }
 mp-hashers = { workspace = true }
 mp-transactions = { workspace = true }
+mp-chain-id = { workspace = true }
 # Starknet dependencies
 blockifier = { workspace = true }
 starknet-ff = { workspace = true }
