+++ crates/runtime/src/pallets.rs
@@ -10,6 +10,7 @@ pub use frame_support::weights::constants::{
 pub use frame_support::weights::{IdentityFee, Weight};
 pub use frame_support::{construct_runtime, parameter_types, StorageValue};
 pub use frame_system::Call as SystemCall;
+pub use mp_chain_id::SN_GOERLI_CHAIN_ID;
 /// Import the StarkNet pallet.
 pub use pallet_starknet;
 pub use pallet_timestamp::Call as TimestampCall;
@@ -22,22 +23,6 @@ pub use sp_runtime::{Perbill, Permill};
 use sp_std::marker::PhantomData;
 
 use crate::*;
-
-/// ChainId for Starknet Goerli testnet
-pub const SN_GOERLI_CHAIN_ID: Felt252Wrapper = Felt252Wrapper(starknet_ff::FieldElement::from_mont([
-    3753493103916128178,
-    18446744073709548950,
-    18446744073709551615,
-    398700013197595345,
-]));
-
-/// ChainId for Starknet Mainnet
-pub const SN_MAIN_CHAIN_ID: Felt252Wrapper = Felt252Wrapper(starknet_ff::FieldElement::from_mont([
-    17696389056366564951,
-    18446744073709551615,
-    18446744073709551615,
-    502562008147966918,
-]));
 
 // Configure FRAME pallets to include in runtime.
